switched structure to inheritance style and refactored code
