+++ lib/address_geocoder/version.rb
@@ -1,0 +1,3 @@
+module AddressGeocoder
+  VERSION = '0.0.1'
+end
